requirements.txt
PyYAML>=3.13 # MIT
zhmcclient>=0.19.0 # Apache-2.0
